feat(driver): add a safety feature to the cmd_vel
In case no cmd_vel has been sent within a small interval, the drone will stop.
This ensures that the controller needs to constantly send the velocity commands
in order to move the drone.

The main reason why this is implemented is if the controller would crash,
and the previous command sent was x=0.5, the drone would not stop.